Инзменен формат входных и выходных данных
* Размер конвертируемой суммы повышен с 9 до 18 символов (long)
* Размер результата конвертации повышен до BigDecimal
* Изменен принцип округления результата конвертации
* Исправлены мелкие баги, мешающие работе нового функционала
* Добавлена обработка ошибки переполнения очереди сервиса получения ставки (code 524)

+++ ClientService/src/main/java/org/emel/ClientService/dto/ConversionInputDataDTO.java
@@ -6,7 +6,7 @@ package org.emel.ClientService.dto;
 public class ConversionInputDataDTO {
     private String fromCurrency;
     private String toCurrency;
-    private int quantity;
+    private long quantity;
     public ConversionInputDataDTO() {
 
     }
@@ -23,10 +23,12 @@ public class ConversionInputDataDTO {
     public void setToCurrency(String toCurrency) {
         this.toCurrency = toCurrency;
     }
-    public int getQuantity() {
+
+    public long getQuantity() {
         return quantity;
     }
-    public void setQuantity(int quantity) {
+
+    public void setQuantity(long quantity) {
         this.quantity = quantity;
     }
 }
